Fix clipboard and its integration with shell
The cliboard was broken, could not copy nor paste. The problem
was that we were not using any particular clipboard and we were
not providing all the arguments necessary to set_text.

Therefore, use Gdk.SELECTION_CLIPBOARD, which is the proper one to
use in this case and pass -1 second argument to set_text [1]. This
also fixes the integration with the sugar shell clipboard.

Tested on Fedora 21 and Ubuntu 14.04.

[1] https://developer.gnome.org/gtk3/3.8/gtk3-Clipboards.html

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>